Python/requirements.txt
pyserial
pymodbus
matplotlib
pillow
python-escpos
